+++ docker/run.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+docker compose -f docker/docker-compose.yml up -d --force-recreate --pull always --remove-orphans strato_acme